+++ src/DropdownLayout/DropdownLayout.js
@@ -129,6 +129,8 @@ class DropdownLayout extends WixComponent {
         break;
       }
 
+      case ' ':
+      case 'Spacebar':
       case 'Enter': {
         if (!this._onSelect(this.state.hovered)) {
           return false;